+++ src/syntax/core/pretty.rs
@@ -1,0 +1,88 @@
+use std::fmt::{Display, Error, Formatter};
+
+use voile_util::tags::{PiSig::*, Plicit};
+
+use super::{Axiom, CaseSplit, Closure, Neutral, Val, ValInfo};
+
+impl Display for Neutral {
+    fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+        use Neutral::*;
+        match self {
+            Var(dbi) => write!(f, "[{}]", dbi),
+            // This might be conflict with other syntax.
+            Ref(dbi) => write!(f, "[|{}|]", dbi),
+            Axi(a) => a.fmt(f),
+            Meta(mi) => write!(f, "?{}", mi),
+            App(fun, a) => {
+                write!(f, "({}", fun)?;
+                for x in a {
+                    write!(f, " {}", x)?;
+                }
+                f.write_str(")")
+            }
+            SplitOn(split, on) => {
+                write!(f, "(case {} of {{ ", on)?;
+                pretty_split(f, &split)?;
+                f.write_str("})")
+            }
+            Fst(p) => write!(f, "({}.1)", p),
+            Snd(p) => write!(f, "({}.2)", p),
+            Lift(levels, p) => write!(f, "(^[{:?}] {})", levels, p),
+        }
+    }
+}
+
+impl Display for Axiom {
+    fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+        use Axiom::*;
+        match self {
+            Postulated(uid) => write!(f, "<{}>", uid),
+            Generated(uid, dbi) => write!(f, "<{} {}>", uid, dbi),
+            Unimplemented(uid, dbi) => write!(f, "[|{} {}|]", uid, dbi),
+            Implicit(uid) => write!(f, "{{{}}}", uid),
+        }
+    }
+}
+
+impl Display for Closure {
+    fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+        use Closure::*;
+        match self {
+            Plain(body) => body.fmt(f),
+            Tree(split) => {
+                for (label, closure) in split {
+                    write!(f, "{} => {}; ", label, closure)?;
+                }
+                Ok(())
+            }
+        }
+    }
+}
+
+impl Display for ValInfo {
+    fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+        write!(f, "{} at {}", self.ast, self.loc)
+    }
+}
+
+impl Display for Val {
+    fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+        match self {
+            Val::Type(l) => write!(f, "set{}", l),
+            Val::Lam(clos) => write!(f, "(\\ {})", clos),
+            Val::Dt(Pi, Plicit::Ex, param_ty, clos) => write!(f, "({} -> {})", param_ty, clos),
+            Val::Dt(Pi, Plicit::Im, param_ty, clos) => write!(f, "({{{}}} -> {})", param_ty, clos),
+            Val::Dt(Sigma, _, param_ty, clos) => write!(f, "({} * {})", param_ty, clos),
+            Val::Pair(fst, snd) => write!(f, "({}, {})", fst, snd),
+            Val::Neut(neut) => neut.fmt(f),
+            Val::Cons(name, a) => write!(f, "(@{} {})", name, a),
+        }
+    }
+}
+
+fn pretty_split(f: &mut Formatter, split: &CaseSplit) -> Result<(), Error> {
+    for (name, closure) in split {
+        write!(f, "{}: \\ {}; ", name, closure)?;
+    }
+    Ok(())
+}