+++ src/syntax/mod.rs
@@ -1,0 +1,2 @@
+/// Core language.
+pub mod core;