+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 0.0.2
 
++ Add core language definition from Voile,
+  with row-polymorphic terms removed (#1)
+
 # 0.0.1
 
 + Create package on crates.io